+++ PetFinderMAUI/PetFinderMAUI/Pages/ProfilePage.xaml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="PetFinderMAUI.Pages.ProfilePage"
+             Title="Account">
+    <ContentPage.Resources>
+        <Style TargetType="BoxView">
+            <Setter Property="HeightRequest" Value="1" />
+            <Setter Property="Color" Value="{StaticResource Gray200}" />
+            <Setter Property="Margin" Value="-20,10" />
+        </Style>
+
+    </ContentPage.Resources>
+    <ScrollView>
+        <VerticalStackLayout Padding="20">
+            <Label Text="my account" FontFamily="PoppinsSemibold" TextColor="{StaticResource Gray300}" />
+            <Label Text="Jesse Smith" FontFamily="PoppinsBold" FontSize="20" />
+            <BoxView /><Label Text="Edit Account Details" /><BoxView />
+            <ContentView HeightRequest="20" />
+
+            <Label Text="APP SETTINGS" FontFamily="PoppinsSemibold" TextColor="{StaticResource Gray300}" />
+            <BoxView />
+            <Label Text="Push Notifications" /><BoxView />
+            <Label Text="Display Settings" /><BoxView />
+            <ContentView HeightRequest="20" />
+
+            <Label Text="SUPPORT" FontFamily="PoppinsSemibold" TextColor="{StaticResource Gray300}" />
+            <BoxView />
+            <Label Text="Frequently Asked Questions" /><BoxView />
+            <Label Text="Report a Bug" /><BoxView />
+            <Label Text="Submit a Question" /><BoxView />
+            <ContentView HeightRequest="40" />
+
+            <BoxView />
+            <Label Text="Log Out">
+                <Label.GestureRecognizers>
+                    <TapGestureRecognizer Tapped="OnLogoutClicked" NumberOfTapsRequired="1" />
+                </Label.GestureRecognizers>
+            </Label>
+
+            <BoxView />
+        </VerticalStackLayout>
+    </ScrollView>
+</ContentPage>
